d/requirements.txt
dnslib==0.9.23
google-generativeai==0.3.2